requirements-dev.txt
argh==0.26.2
blinker==1.4
Bootstrap-Flask==1.0.8
Click==7.0
Faker==0.9.1
Flask==1.0.2
Flask-Avatars==0.2.1
Flask-Dropzone==1.5.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
flask-whooshee==0.6.0
Flask-WTF==0.14.2
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pathtools==0.1.2
Pillow==5.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
PyYAML==3.13
redis==3.3.11
rq==1.1.0
six==1.11.0
SQLAlchemy==1.2.12
text-unidecode==1.2
watchdog==0.9.0
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
gunicorn==19.9.0